fix: issues in comments
