+++ src/components/Calendar/hooks/index.js
@@ -1,6 +1,4 @@
 export { default as useCurrentDateFromValue } from './useCurrentDateFromValue';
 export { default as useRangeFromValue } from './useRangeFromValue';
-export { default as useDaysBuilder } from './useDaysBuilder';
-export { default as useWeeksBuilder } from './useWeeksBuilder';
 export { default as useRangeStartDate } from './useRangeStartDate';
 export { default as useRangeEndDate } from './useRangeEndDate';
