+++ src/components/Calendar/hooks/useRangeEndDate.js
@@ -4,9 +4,9 @@ import { isSameDay, isEmptyRange } from '../helpers';
 export default function useRangeEndDate(date, currentRange) {
     return useMemo(
         () =>
-            !isEmptyRange(currentRange) && currentRange.length >= 2
-                ? isSameDay(date, currentRange[1])
-                : false,
+            isEmptyRange(currentRange) || currentRange.length < 2
+                ? false
+                : isSameDay(date, currentRange[1]),
         [date, currentRange],
     );
 }
